Move MerkleHasher trait into fuzzing.rs
