+++ rollup-interface/src/state_machine/crypto/mod.rs
@@ -1,3 +1,2 @@
 mod simple_hasher;
-pub mod simple_merkle;
 pub use simple_hasher::{NoOpHasher, SimpleHasher};
